% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_colour_scaler.R
\name{lseq}
\alias{lseq}
\title{lseq}
\usage{
lseq(from = 1, to = 1e+05, length_out = 6)
}
\arguments{
\item{from}{start}

\item{to}{end}

\item{length_out}{number of increments}
}
\description{
logarithmic sequence
}
