% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_info_show.R
\name{palette_info_show}
\alias{palette_info_show}
\title{palette_info_show}
\usage{
palette_info_show(
  colour_blind_frieldly = TRUE,
  packages = NULL,
  type = NULL,
  mincolours = 0
)
}
\arguments{
\item{colour_blind_frieldly}{TRUE by default}

\item{packages}{from which packages to display palettes}

\item{type}{qualitative, divergent, sequential, or split palette type}

\item{mincolours}{minimum number of colours needed, default 0}
}
\description{
Visualize palette information
}
