% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_colour_scaler.R
\name{range_check}
\alias{range_check}
\title{range_check}
\usage{
range_check(min, max, var)
}
\arguments{
\item{min}{min}

\item{max}{max}

\item{var}{variable name}
}
\description{
check if a variable with within a range and error if it is not
}
