% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combined_palette_info}
\alias{combined_palette_info}
\title{combined_palette_info}
\format{

}
\usage{
combined_palette_info
}
\description{
Display a summary description of the available palettes
}
\keyword{datasets}
